upgrading more dependencies
Upgrade dependencies in `package.json` and rebuild `yarn.lock`:
- Upgrade `@okta/jwt-verifier` to latest (0.0.12 to 1.0.0)

In the latest version of `okta-jwt-verifier`, `verifyAccessToken` takes a new parameter: `expectedAudience`.

> The expected audience passed to verifyAccessToken() is required, and can be either a string (direct match) or an array strings (the actual aud claim in the token must match one of the strings).

> If you are validating access tokens, you should verify that the aud (audience) claim equals the audience that is configured for your Authorization Server in the Okta Developer Console. -- https://developer.okta.com/code/dotnet/jwt-validation/

https://github.com/okta/okta-oidc-js/blob/6df72c46618afe3f8628dffc38a2386c0cb89179/packages/jwt-verifier/UPGRADING.md

```
Upgrading between versions of @okta/jwt-verifier

Upgrading to the 1.x series

expectedAud is new and required
-------------------------------
The 0.x series verifyAccessToken(tokenString) is replaced by verifyAccessToken(tokenString, expectedAud).

The expectedAud parameter is required, and must match the aud claim within the ticket.

"api://default" is a common value for this claim if not set otherwise.

Additional validations are made (such as the iss claim must match the issuer given to the verifier), but they should always have been true and don't require additional configuration

The clientId is no longer required
----------------------------------
Access Tokens are not required to be bound to a clientId, so the requirement of passing a clientId to the OktaJwtVerifier() constructor has been removed. You can pass the clientId and assert that any cid claim matches by using the "Custom Claims Assertions" as outlined in the README.
```